--- annotation_backup.ini
@@ -1,47 +1,0 @@
-[charging_locations]
-loc0="Charging_0,5.750000,3.400000,-1.566256"
-num=1
-
-[objects]
-num=7
-poly0="Wall_0,1,18.450001:-0.500000,15.400001:-0.500000,15.400001:-1.450000,18.450001:-1.450000"
-poly1="Wall_1,1,18.300001:0.900000,17.000000:0.900000,17.000000:-0.550000,18.300001:-0.550000"
-poly2="Wall_2,1,8.650001:2.950000,9.000000:2.950000,9.000000:3.650000,8.650001:3.650000"
-poly3="Chair_0,1,6.350000:-0.600000,8.250000:-0.600000,8.250000:-1.800000,6.350000:-1.800000"
-poly4="Table_0,1,11.800000:4.550000,6.450000:4.550000,6.450000:3.900000,11.800000:3.900000"
-poly5="Table_1,0,13.200000:1.900000,11.800000:1.900000,12.600000:-0.600000,14.700000:-0.600000,15.550000:0.950000"
-poly6="Chair_1,0,11.200000:0.950000,9.800000:1.300000,9.400001:1.100000,10.150001:-0.450000,11.000000:-0.250000"
-
-[other_locations]
-num=0
-
-[resting_locations]
-loc0="Resting_0,5.550000,1.150000,-0.010309"
-num=1
-
-[serving_0]
-loc0="Serving_0,7.800000,3.250000,-3.086746,"
-loc1="Serving_1,9.000000,-0.550000,1.555664,"
-loc2="Serving_2,10.600000,3.200000,-3.120226,"
-loc3="Serving_3,13.850000,0.000000,-0.005916,"
-loc4="Serving_4,14.600000,3.600000,-1.530721,"
-loc5="Serving_5,16.950001,3.000000,3.112813,"
-loc6="Serving_6,15.850000,0.150000,0.058935,"
-loc7="7777,15.850000,0.150000,0.058935,"
-loc8="888,15.850000,0.150000,0.058935,"
-loc9="9,15.850000,0.150000,0.058935,"
-loc10="10,15.850000,0.150000,0.058935,"
-loc11="11,15.850000,0.150000,0.058935,"
-loc12="12,15.850000,0.150000,0.058935,"
-loc13="13,15.850000,0.150000,0.058935,"
-loc14="14,15.850000,0.150000,0.058935,"
-loc15="Se15rving_6,15.850000,0.150000,0.058935,"
-loc16="Ser16ing_6,15.850000,0.150000,0.058935,"
-loc17="17,15.850000,0.150000,0.058935,"
-loc18="18,15.850000,0.150000,0.058935,"
-loc19="19,15.850000,0.150000,0.058935,"
-name=DEFAULT
-num=20
-
-[serving_locations]
-group=1
