+++ package.json
@@ -34,8 +34,15 @@
     "ws": "^8.17.1"
   },
   "devDependencies": {
+    "glob": "^11.0.0",
+    "gulp": "^5.0.0",
+    "gulp-javascript-obfuscator": "^1.1.6",
+    "gulp-uglify-es": "^3.0.0",
     "node-pre-gyp": "^0.17.0",
-    "nodemon": "^3.1.3"
+    "nodemon": "^3.1.3",
+    "webpack": "^5.96.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-obfuscator": "^3.5.1"
   },
   "optionalDependencies": {
     "bufferutil": "^4.0.8",
